+++ apps/docs/app/page.tsx
@@ -1,6 +1,6 @@
 import Image, { type ImageProps } from "next/image";
-import { Button } from "../../../packages/react/button/src/button";
 import styles from "./page.module.css";
+import { Button } from "@everest-ui/react-button";
 
 type Props = Omit<ImageProps, "src"> & {
   srcLight: string;