+++ packages/react/button/postcss.config.js
@@ -1,9 +1,0 @@
-import PostcssModulesPlugin from "postcss-modules";
-
-export default {
-  plugins: [
-    PostcssModulesPlugin({
-      generateScopedName: "[name]__[local]___[hash:base64:5]",
-    }),
-  ],
-};