+++ packages/react/button/rollup.config.js
@@ -1,0 +1,5 @@
+import Rollup from "@everest-ui/build-config/rollup";
+
+export default {
+  ...Rollup,
+};