+++ apps/docs/CHANGELOG.md
@@ -1,5 +1,23 @@
 # docs
 
+## 2.0.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @everest-ui/react-button@1.2.4
+
+## 2.0.0
+
+### Major Changes
+
+- Add all packages
+
+### Patch Changes
+
+- Updated dependencies
+  - @everest-ui/react-button@2.0.0
+
 ## 1.0.0
 
 ### Major Changes