+++ packages/typescript-config/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @repo/typescript-config
 
+## 2.0.0
+
+### Major Changes
+
+- Add all packages
+
 ## 1.0.0
 
 ### Major Changes